Refactor form components: rename files and update imports
