+++ market/src/components/ui/FormError.tsx
@@ -1,0 +1,13 @@
+import type { ReactNode } from "react";
+
+interface FormErrorProps {
+  message: string | ReactNode;
+  className?: string;
+}
+
+export function FormError({ message, className = "" }: FormErrorProps) {
+  if (!message) return null;
+  return (
+    <div className={`form-error ${className}`}>{message}</div>
+  );
+}
