+++ market/src/components/ui/FormField.tsx
@@ -1,0 +1,21 @@
+import type { ReactNode } from "react";
+
+interface FormFieldProps {
+  label: string;
+  htmlFor: string;
+  children: ReactNode;
+  error?: string | null;
+  className?: string;
+}
+
+export function FormField({ label, htmlFor, children, error, className = "" }: FormFieldProps) {
+  return (
+    <div className={`space-y-1 ${className}`}>
+      <label htmlFor={htmlFor} className="block text-sm font-medium text-gray-700">
+        {label}
+      </label>
+      {children}
+      {error && <p className="form-error">{error}</p>}
+    </div>
+  );
+}
